[FIX] MarketService 파일 2개 중 하나 삭제
