+++ src/main/java/com/kanghwang/khholdings/domain/order/service/OrderService.java
@@ -39,7 +39,7 @@ public class OrderService {
 
 	// 매수/매도 주문
 	@Transactional
-	public boolean placeOrder(OrderRequestDTO orderDTO) {
+	public void placeOrder(OrderRequestDTO orderDTO) {
 
 		// 1. Snowflake ID를 사용하여 주문 번호 생성
 		Long orderId = snowflakeIdGenerator.nextId();
@@ -71,8 +71,6 @@ public class OrderService {
 				orderRedisService.processOrder(orderDTO);
 			}
 		});
-
-		return true;
 	}
 
 	// 주문 취소
